Spring Pads now launch characters
~BUG~
The character that weighs the pad down more appears to be lighter? Not sure what's causing this...

Cleaner implementation would have a blueprint inside character that takes the upwards force of the pad and converts it into a launch, currently all the launch code takes place within the spring pad blueprints which is a bit messy.